Fase 1: Extracted utility functions and types for time-clock
