+++ app/reloj-checador/page.tsx
@@ -2,7 +2,7 @@
 
 import { Suspense } from 'react';
 import { useSearchParams } from 'next/navigation';
-import TimeClock from '@/app/components/time-clock'; // Ajusta la ruta si es necesario
+import TimeClock from '@/app/components/time-clock-view'; // Ajusta la ruta si es necesario
 import { Loader2 } from 'lucide-react';
 
 // Componente interno que usa useSearchParams
